fix: uint8_t overflow in altar cost formula
10*(t+1)*t reaches 300 at t=5, overflowing uint8_t (max 255).
Change altar cost and internal calculation to uint16_t.

+++ inc/event/event_manager.h
@@ -23,7 +23,7 @@ private:
 	uint8_t altar_times_;  // 全局祭坛使用次数
 
 	static uint8_t countMonsters(uint8_t floor);
-	uint8_t getAltarCost() const;
+	uint16_t getAltarCost() const;
 
 	void executeEvent(const Event& ev, Player& player);
 	void executeAction(const EventAction& act, Player& player, uint8_t ev_floor);
